Bruno

+++ PechinchaMarket/Controllers/ManagerController.cs
@@ -109,10 +109,14 @@ namespace PechinchaMarket.Controllers
             if (comerciante != null)
             {
                 comerciante.isApproved = true;
-                if (utilizador != null)
-                {
-                    await SendEmailAsync(utilizador.Email, "Seu registo foi aceite",
-              "Estamos felizes em informar que seu registo como comerciante na plataforma PechinchaMarket foi aceite");
+
+                if(utilizador != null) {
+                    utilizador.EmailConfirmed = true;
+                    EmailSender emailSender = new EmailSender();
+                    await emailSender.SendEmail("Seu cadastro foi aceito",utilizador.Email,comerciante.Name, "Estamos felizes em informar que seu registo como comerciante na plataforma PechinchaMarket foi aceito",_context);
+                   
+            
+
                 }
             }
             await _context.SaveChangesAsync();
@@ -158,10 +162,20 @@ namespace PechinchaMarket.Controllers
 
             if (comerciante != null)
             {
-                await SendEmailAsync(utilizador.Email, "Seu registo foi negado",
-              "Lamentamos informar que seu registo como comerciante na plataforma PechinchaMarket não foi aceite");
-                _context.Comerciante.Remove(comerciante);
-                _context.Users.Remove(utilizadorId);
+
+                var utilizador = await _context.Users.FirstOrDefaultAsync(m => m.Id == comerciante.UserId);
+        
+             
+                if (utilizador != null)
+                {
+                    EmailSender emailSender = new EmailSender();
+                    await emailSender.SendEmail("Seu cadastro foi recusado", utilizador.Email, comerciante.Name, "Lamentamos informar que seu registo como comerciante na plataforma PechinchaMarket não foi aceito",_context);
+                    _context.Comerciante.Remove(comerciante);
+                    _context.Users.Remove(utilizador);
+                }
+            
+                
+
 
             }
             await _context.SaveChangesAsync();
